requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
sqlalchemy
aiosqlite
google-generativeai
python-docx
tenacity
httpx
python-dotenv
pypdf2